Déplacement des fonctions dans leur components
Suite à la division du code grade en plus petits components, il faut également déplacer les fonctions qui sont liées à chaque bout de code
