+++ frontend/src/components/ModalConfirmationCancel.js
@@ -3,6 +3,7 @@ import '../css/Grades.css';
 
 const ModalConfirmationCancel = ({activateButton, deleteModification}) => {
 
+    
     return (
         <div id="confirmationCancelModal" className="modal fade" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="confirmationCancelModalLabel" aria-hidden="true">
             <div className="modal-dialog modal-lg modal-dialog-centered modal-dialog-scrollable">
